requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-celery-results==1.1.2
djangorestframework==3.10.2
idna==2.8
importlib-metadata==0.19
kombu==4.6.4
python-http-client==3.1.0
pytz==2019.2
requests==2.22.0
sendgrid==6.0.5
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
zipp==0.5.2
